+++ .github/workflows/ci.yml
@@ -23,4 +23,4 @@ jobs:
     - name: Lint with flake8
       run: |
         pip install flake8
-        flake8 ./handlers
+        flake8 ./handlers ./sources
