+++ package.json
@@ -9,7 +9,7 @@
   "author": "Antosik",
   "license": "ISC",
   "dependencies": {
-    "serverless": "^1.71.3",
+    "serverless": "^1.72.0",
     "serverless-python-requirements": "^5.0.1"
   }
 }
